requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
cycler==0.10.0
decorator==4.4.2
gast==0.4.0
google-pasta==0.2.0
grpcio==1.34.0
h5py==3.1.0
imageio==2.9.0
importlib-metadata==3.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
Pillow==8.0.1
progressbar2==3.53.1
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.4.0
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tifffile==2020.9.3
Werkzeug==1.0.1
wrapt==1.12.1
zhusuan @ file:///C:/Users/hector/Documents/GitHub/zhusuan-mnist/zhusuan-master
zipp==3.4.0
